GSCHED-702: common Program class for OCS/GPP
